--- src/Contracts/ReceiptInterface.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Shetabit\Payment\Contracts;
-
-interface ReceiptInterface
-{
-    /**
-     * Retrieve tracking code.
-     *
-     * @return string|integer
-     */
-    public function getTrackingCode();
-}
